+++ external/admob/src/commonMain/kotlin/com/whatever/caramel/external/admob/banner/CaramelAdBanner.kt
@@ -1,9 +1,10 @@
-package com.whatever.caramel.core.ui.admob
+package com.whatever.caramel.external.admob.banner
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.staticCompositionLocalOf
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -12,6 +13,11 @@ import caramel.core.designsystem.generated.resources.Res
 import caramel.core.designsystem.generated.resources.ad_banner
 import com.whatever.caramel.core.designsystem.themes.CaramelTheme
 import org.jetbrains.compose.resources.stringResource
+
+val LocalHomeGoogleAdBanner =
+    staticCompositionLocalOf<GoogleAdBannerState> {
+        error("No HomeGoogleAdBanner provided")
+    }
 
 @Composable
 fun CaramelGoogleAdBanner(
