+++ external/admob/src/commonMain/kotlin/com/whatever/caramel/external/admob/banner/GoogleAdBannerView.kt
@@ -1,16 +1,10 @@
 @file:Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
 
-package com.whatever.caramel.core.ui.admob
+package com.whatever.caramel.external.admob.banner
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.Stable
-import androidx.compose.runtime.staticCompositionLocalOf
 import androidx.compose.ui.Modifier
-
-val LocalHomeGoogleAdBanner =
-    staticCompositionLocalOf<GoogleAdBannerState> {
-        error("No HomeGoogleAdBanner provided")
-    }
 
 /**
  * 광고 Unit Id 모음
@@ -74,6 +68,6 @@ sealed interface GoogleAdBannerType {
  */
 @Composable
 expect fun rememberGoogleAdBannerState(
-    adUnitId: String,
-    bannerType: GoogleAdBannerType,
+    adUnitId: String = GoogleAdUnitIds.TEST_BANNER,
+    bannerType: GoogleAdBannerType = GoogleAdBannerType.AnchoredAdaptive(),
 ): GoogleAdBannerState
