+++ settings.gradle.kts
@@ -65,3 +65,4 @@ include(":external:inapp-review")
 include(":external:deeplink")
 include(":external:firebase-messaging")
 include(":external:analytics")
+include(":external:admob")
